backend/requirements.txt
fastapi
uvicorn
pymongo
networkx
openai
GitPython
python-dotenv
pydantic
